+++ src/components/Tabs/types.ts
