+++ src/shared/utils/isEqualKeyValue.ts
@@ -1,8 +1,8 @@
 import { ControlProps } from '../types';
 
 function isEqualKeyValue(
-  a: ControlProps['keyValue'],
-  b: ControlProps['keyValue'],
+  a?: ControlProps['keyValue'],
+  b?: ControlProps['keyValue'],
 ) {
   return a === b;
 }
